pyproject.toml
[project]
name = "sqlagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.7",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "streamlit>=1.50.0",
]
